HELPS-123: RG - splitting tests into its own group
