+++ ldk/javascript/examples/self-test-loop/src/tests/user/index.ts
@@ -1,0 +1,13 @@
+import { user } from '@oliveai/ldk';
+
+export const userJWTTest = (): Promise<boolean> =>
+  new Promise((resolve, reject) => {
+    user.jwt().then((token) => {
+      if (token) {
+        console.debug('jwt', token);
+        resolve(true);
+      } else {
+        reject(new Error('JWT should not have been empty'));
+      }
+    });
+  });
